arch: no import cycles left, at either level
kcn: finish the decoupling rather than leaving what is unsatisfied. Both module
cycles are gone, so KNOWN_MODULE_CYCLES joins KNOWN_PACKAGE_CYCLES at empty and
the assertion says "none" instead of "at most two".

tools <-> tools.context_tools I had called idiomatic and left. It is a common
shape — every tool module imports BaseTool from the package, the package imports
the tool modules back to build the registry, and `build_registry` defers its
import so it works — but common is not the same as acyclic. tools.base now holds
the three definitions, which depend on nothing, so both sides sit above them.

decision.ledger <-> decision.record was the same trick one layer down: `record`
appends through `ledger`, and `ledger` reached back into `record` to validate one
row type from inside a function body. The schema and its four vocabularies had no
dependency on either module, so they moved to decision.mind_record and the
deferred import became an ordinary one.

A function-level import is not a fix for a cycle. It is a cycle that starts
working, and both of these were that.

The parallelism claim in #816 turned out to be wrong, and the correction is
recorded where the claim was. Emptying the writer list was supposed to unlock
`-n auto` and cut the pytest step from 89s to ~30s. The session rebuild is now
behind a file lock so four workers cannot build the same four files at once, and
`-n 4` passes — at 175/182/191s against ~195s serial. Five to ten percent. The
twelve slowest tests are ~121s of a 208s run and every one is subprocess-bound,
so the suite is deep in a few places rather than wide. One of four parallel runs
also failed in a way that would not reproduce. xdist stays off; a few percent
does not buy a flake class. The isolation was worth doing regardless — it fixed
a real order-dependent failure — and the lock is correct whether or not anyone
runs workers.

Closes #816
