+++ src/clawock/decision/record.py
@@ -20,6 +20,13 @@ Usage:
 """
 from __future__ import annotations
 
+from clawock.decision.mind_record import (  # noqa: F401  (re-export)
+    ACTIONS,
+    DRIVEN_BY,
+    EMOTIONS,
+    SOURCES,
+    validate_mind_record,
+)
 import argparse
 import json
 import sys
@@ -28,14 +35,6 @@ from pathlib import Path
 
 from clawock.decision import ledger as decision_v2
 
-ACTIONS = {"buy", "add", "trim", "sell", "hold", "watch", "reject", "abstain"}
-DRIVEN_BY = {"technical", "fundamental", "sentiment", "mixed"}
-EMOTIONS = {"calm", "fomo", "revenge", "averaging_down", "fear", "euphoria", "mixed"}
-# Which harness produced the verdict. `conversation` is the DSH default and
-# the historical value; other harnesses pass their own so the ledger can say
-# where a mind record came from. `brief` is the desk's daily plan writer and
-# is not a valid record() value — it is produced by the brief postflight.
-SOURCES = {"conversation", "openclaw", "claude", "codex", "cli"}
 MIND_SCHEMA_VERSION = 0
 
 
@@ -43,33 +42,6 @@ def _now_iso() -> str:
     return datetime.now(timezone.utc).astimezone().isoformat(timespec="seconds")
 
 
-def validate_mind_record(record: dict) -> list[str]:
-    """Return human-readable issues; empty means the record is appendable."""
-    issues: list[str] = []
-    subject = record.get("subject") or {}
-    for field in ("ticker", "market", "currency"):
-        if not isinstance(subject.get(field), str) or not subject.get(field, "").strip():
-            issues.append(f"subject.{field} is required")
-    if record.get("action") not in ACTIONS:
-        issues.append(f"action must be one of {sorted(ACTIONS)}")
-    if record.get("source") not in SOURCES:
-        issues.append(f"source must be one of {sorted(SOURCES)}")
-    confidence = record.get("confidence")
-    if isinstance(confidence, bool) or not isinstance(confidence, (int, float)) or not 0 <= confidence <= 1:
-        issues.append("confidence must be between 0 and 1")
-    if record.get("driven_by") not in DRIVEN_BY:
-        issues.append(f"driven_by must be one of {sorted(DRIVEN_BY)}")
-    mind = record.get("mind") or {}
-    for side in ("bull", "bear"):
-        case = mind.get(side) or {}
-        if not isinstance(case.get("summary"), str) or not case.get("summary", "").strip():
-            issues.append(f"mind.{side}.summary is required (opposing case is mandatory)")
-    if not isinstance(mind.get("invalidation"), list) or not mind.get("invalidation"):
-        issues.append("mind.invalidation must be a non-empty list of observable conditions")
-    emotion = record.get("emotion") or {}
-    if emotion.get("pressure") not in EMOTIONS:
-        issues.append(f"emotion.pressure must be one of {sorted(EMOTIONS)}")
-    return issues
 
 
 def build_record(args) -> dict:
