+++ src/clawock/tools/context_tools.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 from clawock.context import brief as brief_context
 from clawock.decision import packet as brief_decision_packet
-from clawock.tools import BaseTool, ToolError
+from clawock.tools.base import BaseTool, ToolError
 
 SECTIONS = ("facts", "technical", "quant", "sentiment", "evidence", "risk",
             "constraints")
